+++ aws/kms/examples/complete/main.tf
@@ -1,6 +1,6 @@
 module "kms" {
   source                   = "../../"
-  name                     = "example"
+  alias_name               = ["example"]
   description              = "example"
   deletion_window_in_days  = 7
   key_usage                = "ENCRYPT_DECRYPT"