+++ aws/kms/CHANGELOG.md
@@ -1,6 +1,12 @@
 # Changelog
 
 All notable changes to this project will be documented in this file.
+## [0.1.9]() (2024-12-09)
+### Features
+* Update README.md
+* Change name -> alias_name and make it into a `list(string)`
+* Update examples, outputs and description of some variables.
+
 ## [0.1.4]() (2024-12-05)
 ### Features
 * Update terraform version constraint from `~> 1.9.8` to `>= 1.9.8`